Update send-and-transfer.sol
I removed this piece of code from line 39.

  require(msg.value >= 1 ether);
It does not necessary in the code, because we defined the costs modifier in line 25-26.

-This bug was discovered by Jon_m -

Bug fix:

line98: uint toTransfer = balance; → It was missing the uint command.
line100: msg.sender.transfer(toTransfer); → Like you mentioned before, this is how it is supposed to be
The code is failing in the original case, if you try to withdraw all your funds from the contract. That is gonna wipe itself and that is not gonna return for the contract owner. In the previous line we defined balance = 0;, after there is no funds to send to the owner.